WIP: move recommendation API out of collection
