+++ lib/collection/tests/multi_vec_test.rs
@@ -218,15 +218,10 @@ async fn test_multi_vec_with_shards(shard_number: u32) {
         .recommend_by(
             RecommendRequest {
                 positive: vec![6.into()],
-                negative: vec![],
                 with_payload: Some(WithPayloadInterface::Bool(false)),
                 with_vector: Some(WithVector::Selector(vec![VEC_NAME2.to_string()])),
-                score_threshold: None,
                 limit: 10,
-                offset: 0,
-                filter: None,
-                params: None,
-                using: None,
+                ..Default::default()
             },
             &Handle::current(),
             None,
@@ -245,15 +240,11 @@ async fn test_multi_vec_with_shards(shard_number: u32) {
         .recommend_by(
             RecommendRequest {
                 positive: vec![6.into()],
-                negative: vec![],
                 with_payload: Some(WithPayloadInterface::Bool(false)),
                 with_vector: Some(WithVector::Selector(vec![VEC_NAME2.to_string()])),
-                score_threshold: None,
                 limit: 10,
-                offset: 0,
-                filter: None,
-                params: None,
                 using: Some(VEC_NAME1.to_string().into()),
+                ..Default::default()
             },
             &Handle::current(),
             None,
